requirements.txt
langchain==0.2.14
PyPDF2==3.0.1
python-dotenv==1.0.1
streamlit==1.37.1
